Add standard schema types
https://standardschema.dev/
