+++ src/types.ts
@@ -1,4 +1,5 @@
 import type { Event, Effect, Store, Unit, Subscription } from 'effector'
+import type { StandardSchemaV1 } from './standard-schema'
 
 export interface Adapter<State> {
   get( //
@@ -33,6 +34,7 @@ export interface StorageAdapterFactory<AdapterConfig> {
 
 export type Contract<Data> =
   | ((raw: unknown) => raw is Data)
+  | StandardSchemaV1<unknown, Data>
   | {
       isData: (raw: unknown) => raw is Data
       getErrorMessages: (raw: unknown) => string[]
